pyproject.toml
[project]
name = "apiautomationplayground"
version = "1.1.6"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "django-allauth[headless]>=65.14.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "pillow>=12.1.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
prod = [
    "mysqlclient>=2.2.8",
]
